+++ src/image/_example/placeholder.jsx
@@ -46,7 +46,7 @@ export default function PlaceholderImage() {
           <Space direction="vertical">
             默认错误
             <Image
-              src=""
+              src="error url"
               key={errorCount}
               style={{ width: 284, height: 160 }}
               onError={(e) => {
@@ -59,7 +59,7 @@ export default function PlaceholderImage() {
           </Space>
           <Space direction="vertical">
             自定义错误
-            <Image key={errorCount} style={{ width: 284, height: 160 }} error={<ImageErrorIcon size={24} />} />
+            <Image src="error url" key={errorCount} style={{ width: 284, height: 160 }} error={<ImageErrorIcon size={24} />} />
           </Space>
         </Space>
       </div>