+++ src/table/defaultProps.ts
@@ -2,7 +2,7 @@
  * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
  * */
 
-import { TdBaseTableProps, TdPrimaryTableProps } from './type';
+import { TdBaseTableProps, TdPrimaryTableProps, TdEnhancedTableProps } from './type';
 
 export const baseTableDefaultProps: TdBaseTableProps = {
   bordered: false,
@@ -14,6 +14,7 @@ export const baseTableDefaultProps: TdBaseTableProps = {
   footerAffixedBottom: false,
   headerAffixedTop: false,
   hover: false,
+  lazyLoad: false,
   loading: undefined,
   resizable: false,
   rowKey: 'id',
@@ -46,3 +47,5 @@ export const primaryTableDefaultProps: Pick<
   showSortColumnBgColor: false,
   sortOnRowDraggable: false,
 };
+
+export const enhancedTableDefaultProps: TdEnhancedTableProps = { defaultExpandedTreeNodes: [], rowKey: undefined };