+++ src/table/type.ts
@@ -24,6 +24,7 @@ import {
   OptionData,
   SizeEnum,
   ClassName,
+  Styles,
   AttachNode,
   HTMLElementAttributes,
   ComponentType,
@@ -125,6 +126,11 @@ export interface TdBaseTableProps<T extends TableRowData = TableRowData> {
    */
   lastFullRow?: TNode;
   /**
+   * 是否启用整个表格元素的懒加载，当页面滚动到可视区域后再渲染表格。注意和表格内部行滚动懒加载的区别，内部行滚动无论表格是否在可视区域都会默认渲染第一屏的行元素
+   * @default false
+   */
+  lazyLoad?: boolean;
+  /**
    * 加载中状态。值为 `true` 会显示默认加载中样式，可以通过 Function 和 插槽 自定义加载状态呈现内容和样式。值为 `false` 则会取消加载状态
    */
   loading?: TNode;
@@ -145,7 +151,7 @@ export interface TdBaseTableProps<T extends TableRowData = TableRowData> {
    */
   paginationAffixedBottom?: boolean | Partial<AffixProps>;
   /**
-   * 是否允许调整列宽。如果想要配置宽度可调整的最小值和最大值，请使用 `column.resize`，示例：`columns: [{ resize: { minWidth: 120, maxWidth: 300 } }]`。<br/> 默认规则：因列宽超出存在横向滚动条时，列宽调整仅影响当前列宽和总列宽；表格列较少没有横向滚动条时，列宽调整表现为自身宽度和相邻宽度变化
+   * 是否允许调整列宽，设置 `tableLayout=fixed` 效果更友好，此时不允许通过 CSS 设置 `table`元素宽度，也不允许设置 `tableContentWidth`。一般不建议在列宽调整场景使用 `tableLayout: auto`。如果想要配置宽度可调整的最小值和最大值，请使用 `column.resize`，示例：`columns: [{ resize: { minWidth: 120, maxWidth: 300 } }]`。<br/> 默认规则：因列宽超出存在横向滚动条时，列宽调整仅影响当前列宽和总列宽；表格列较少没有横向滚动条时，列宽调整表现为自身宽度和相邻宽度变化
    * @default false
    */
   resizable?: boolean;
@@ -195,7 +201,7 @@ export interface TdBaseTableProps<T extends TableRowData = TableRowData> {
    */
   tableContentWidth?: string;
   /**
-   * 表格布局方式
+   * 表格布局方式，`<table>` 元素原生属性。[MDN](https://developer.mozilla.org/en-US/docs/Web/CSS/table-layout)。注意，在列宽调整下场景只能使用 `fixed` 模式
    * @default fixed
    */
   tableLayout?: 'auto' | 'fixed';
@@ -312,7 +318,7 @@ export interface BaseTableCol<T extends TableRowData = TableRowData> {
    */
   colspan?: number;
   /**
-   * 单元格和表头内容超出时，是否显示省略号。如果仅希望单元格超出省略，可设置 `ellipsisTitle = false`。<br/> 值为 `true`，则超出省略浮层默认显示单元格内容；<br/>值类型为 `Function` 则自定义超出省略浮中层显示的内容；<br/>值类型为 `Object`，则自动透传属性到 Tooltip 组件，可用于调整浮层背景色和方向等特性。<br/> 同时透传 Tooltip 属性和自定义浮层内容，请使用 `{ props: { theme: 'light' }, content: () => 'something' }`
+   * 单元格和表头内容超出时，是否显示省略号。如果仅希望单元格超出省略，可设置 `ellipsisTitle = false`。<br/> 值为 `true`，则超出省略浮层默认显示单元格内容；<br/>值类型为 `Function` 则自定义超出省略浮中层显示的内容；<br/>值类型为 `Object`，则自动透传属性到 Tooltip 组件，可用于调整浮层背景色和方向等特性。<br/> 同时透传 Tooltip 属性和自定义浮层内容，请使用 `{ props: { theme: 'light' }, content: () => 'something' }`。<br /> 请注意单元格超出省略的两个基本点：1. 内容元素是内联元素或样式（自定义单元格内容时需特别注意）；2. 内容超出父元素
    * @default false
    */
   ellipsis?:
@@ -552,7 +558,7 @@ export interface TdPrimaryTableProps<T extends TableRowData = TableRowData>
   /**
    * 过滤参数发生变化时触发，泛型 T 指表格数据类型
    */
-  onFilterChange?: (filterValue: FilterValue, context: { col?: PrimaryTableCol<T> }) => void;
+  onFilterChange?: (filterValue: FilterValue, context: TableFilterChangeContext<T>) => void;
   /**
    * 行编辑时触发
    */
@@ -649,6 +655,16 @@ export interface TdEnhancedTableProps<T extends TableRowData = TableRowData> ext
    */
   beforeDragSort?: (context: DragSortContext<T>) => boolean;
   /**
+   * 展开的树形节点。非必须。在需要自由控制展开的树形节点时使用。其他场景无需设置，表格组件有内置展开逻辑
+   * @default []
+   */
+  expandedTreeNodes?: Array<string | number>;
+  /**
+   * 展开的树形节点。非必须。在需要自由控制展开的树形节点时使用。其他场景无需设置，表格组件有内置展开逻辑，非受控属性
+   * @default []
+   */
+  defaultExpandedTreeNodes?: Array<string | number>;
+  /**
    * 树形结构相关配置。具体属性文档查看 `TableTreeConfig` 相关描述
    */
   tree?: TableTreeConfig;
@@ -661,7 +677,15 @@ export interface TdEnhancedTableProps<T extends TableRowData = TableRowData> ext
    */
   onAbnormalDragSort?: (context: TableAbnormalDragSortContext<T>) => void;
   /**
-   * 树形结构，用户操作引起节点展开或收起时触发，代码操作不会触发
+   * 树形结构，展开的树节点发生变化时触发，泛型 T 指表格数据类型
+   */
+  onExpandedTreeNodesChange?: (
+    expandedTreeNodes: Array<string | number>,
+    options: TableTreeNodeExpandOptions<T>,
+  ) => void;
+  /**
+   * 树形结构，用户操作引起节点展开或收起时触发。请更为使用 `onExpandedTreeNodesChange`
+   * @deprecated
    */
   onTreeExpandChange?: (context: TableTreeExpandChangeContext<T>) => void;
 }
@@ -704,6 +728,10 @@ export interface EnhancedTableInstanceFunctions<T extends TableRowData = TableRo
    * 树形结构中，移除指定节点
    */
   remove: (key: TableRowValue) => void;
+  /**
+   * 树形结构中，移除指定节点的所有子节点
+   */
+  removeChildren: (key: TableRowValue) => void;
   /**
    * 重置或更新整个表格数据
    */
@@ -765,6 +793,15 @@ export interface TableRowState<T extends TableRowData = TableRowData> {
 
 export interface TableColumnFilter {
   /**
+   * 用于透传筛选器属性到自定义组件 `component`，HTML 原生属性
+   */
+  attrs?: HTMLElementAttributes;
+  /**
+   * 透传类名到自定义组件 `component`
+   * @default ''
+   */
+  classNames?: ClassName;
+  /**
    * 用于自定义筛选器，只要保证自定义筛选器包含 value 属性 和 change 事件，即可像内置筛选器一样正常使用。示例：`component: DatePicker`
    */
   component?: ComponentType;
@@ -781,7 +818,7 @@ export interface TableColumnFilter {
    */
   popupProps?: PopupProps;
   /**
-   * 用于透传筛选器属性，可以对筛选器进行任何原组件支持的属性配置
+   * 用于透传筛选器属性到自定义组件 `component`，可以对筛选器进行任何原组件支持的属性配置
    */
   props?: FilterProps;
   /**
@@ -794,7 +831,11 @@ export interface TableColumnFilter {
    */
   showConfirmAndReset?: boolean;
   /**
-   * 用于设置筛选器类型：单选按钮筛选器、复选框筛选器、输入框筛选器
+   * 透传内联样式到自定义组件 `component`
+   */
+  style?: Styles;
+  /**
+   * 用于设置筛选器类型：单选按钮筛选器、复选框筛选器、输入框筛选器。更多复杂组件，请更为使用 `component` 自定义任意组件
    * @default ''
    */
   type?: FilterType;
@@ -849,6 +890,11 @@ export interface TableEditableCellConfig<T extends TableRowData = TableRowData> 
    */
   defaultEditable?: boolean;
   /**
+   * 设置当前列的单元格始终保持为编辑态
+   * @default false
+   */
+  keepEditMode?: boolean;
+  /**
    * 透传给编辑组件的事件
    */
   on?: (context: TableEditableCellPropsParams<T>) => { [eventName: string]: Function };
@@ -879,7 +925,7 @@ export interface TableEditableCellConfig<T extends TableRowData = TableRowData> 
 export interface TableTreeConfig {
   /**
    * 表示树形结构的行选中（多选），父子行选中是否独立
-   * @default true
+   * @default false
    */
   checkStrictly?: boolean;
   /**
@@ -1050,6 +1096,11 @@ export interface ExpandOptions<T> {
   currentRowData: T;
 }
 
+export interface TableFilterChangeContext<T> {
+  col?: PrimaryTableCol<T>;
+  trigger: 'filter-change' | 'confirm' | 'reset' | 'clear';
+}
+
 export type PrimaryTableRowEditContext<T> = PrimaryTableCellParams<T> & { value: any; editedRow: T };
 
 export type PrimaryTableRowValidateContext<T> = { result: TableRowValidateResult<T>[]; trigger: TableValidateTrigger };
@@ -1103,6 +1154,14 @@ export interface TableAbnormalDragSortContext<T> {
   reason: string;
 }
 
+export interface TableTreeNodeExpandOptions<T> {
+  row: T;
+  rowIndex: number;
+  rowState: TableRowState<T>;
+  type: 'fold' | 'expand';
+  trigger?: 'expand-fold-icon' | 'row-click' | 'default-expand-all' | 'expand-all' | 'fold-all';
+}
+
 export interface TableTreeExpandChangeContext<T> {
   row: T;
   rowIndex: number;
@@ -1131,6 +1190,7 @@ export type TableEditableCellProps<T> =
 
 export interface TableEditableCellPropsParams<T> extends PrimaryTableCellParams<T> {
   editedRow: T;
+  updateEditedCellValue: (val: any) => void;
 }
 
 export interface TablePlainObject {