+++ src/tree/hooks/useTreeVirtualScroll.ts
@@ -1,8 +1,7 @@
 import { useMemo, useEffect, CSSProperties } from 'react';
 import useVirtualScroll from '../../hooks/useVirtualScroll';
 import TreeNode from '../../_common/js/tree/tree-node';
-
-import type { TScroll } from '../../common';
+import { TScroll } from '../../common';
 
 export default function useTreeVirtualScroll({
   treeRef,
@@ -39,6 +38,7 @@ export default function useTreeVirtualScroll({
     scrollHeight = null,
     translateY = null,
     handleRowMounted = null,
+    scrollToElement,
   } = useVirtualScroll(treeRef, {
     data: data || [],
     scroll: scrollParams,
@@ -100,5 +100,6 @@ export default function useTreeVirtualScroll({
     isVirtual,
     cursorStyle,
     treeNodeStyle,
+    scrollToElement,
   };
 }